Fleshed out BinaryTree functionality, included some TreeMap code
The BinaryTree class now has recursive pre/in/post-order traversal
methods and a recursive insertion method.

Main class now has code to make a TreeMap with a number of key-value
entries, as well as a Set to put those entries in and an Iterator with
which to print out its ordered key-value pairs.