Overloaded LinkedList add method, improved Stack efficiency
Overloaded the add method in the LinkedList class to also be able to
take an index to place our new node into, instead of just always placing
the node at the end of the list.  Stack now uses the overloaded method
and pushes nodes on the stack by placing them at, and popping them from,
the front of the list.